Big changes

+++ EventsServer/Server/Controllers/EventsController.cs
@@ -10,7 +10,6 @@ using Server.ActionFilters;
 
 namespace Server.Controllers
 {
-    [Authorize]
     [ApiController]
     [ApiVersion("1.0")]
     [Route("modsen/events")]
@@ -18,10 +17,13 @@ namespace Server.Controllers
     {
         private readonly IRepositoryManager _repository;
         private readonly IMapper _mapper;
-        public EventsController(IRepositoryManager repository, IMapper mapper)
+        private readonly ILoggerManager _logger;
+
+        public EventsController(IRepositoryManager repository, IMapper mapper, ILoggerManager logger)
         {
             _repository = repository;
             _mapper = mapper;
+            _logger = logger;
         }
         /// <summary>
         /// Return all acces requests for this url
@@ -46,7 +48,6 @@ namespace Server.Controllers
         /// <response code="200">Returns the list of all events</response>
         /// <response code="400">Request parameters are not valid</response>
         /// <response code="500">Server error</response>
-
         [HttpGet]
         [HttpHead]
         [ProducesResponseType(200)]
@@ -65,21 +66,18 @@ namespace Server.Controllers
 
             return Ok(eventsDto);
         }
+
         /// <summary>
         /// Return one event by id
         /// </summary>
         /// <param name="id">Event id</param>
         /// <returns>Return one event</returns>
         /// <response code="200">Returns one event</response>
-        /// <response code="401">You are not authorized</response>
-        /// <response code="403">You have no rules</response>
         /// <response code="404">Event with this id not found</response>
         /// <response code="500">Server error</response>
         [HttpGet("{id}", Name = "GetEventById")]
         [ServiceFilter(typeof(ValidateEventExistsAttribute))]
         [ProducesResponseType(200)]
-        [ProducesResponseType(401)]
-        [ProducesResponseType(403)]
         [ProducesResponseType(404)]
         [ProducesResponseType(500)]
         public IActionResult GetEventById(Guid id)
@@ -89,6 +87,38 @@ namespace Server.Controllers
             var eventDto = _mapper.Map<EventDto>(singleEvent);
             return Ok(eventDto);
         }
+
+        /// <summary>
+        /// Return all events for current user
+        /// </summary>
+        /// <param name="eventParameters">Filter parameters</param>
+        /// <returns>The events list</returns>
+        /// <response code="200">Returns the list of all events</response>
+        /// <response code="400">Request parameters are not valid</response>
+        /// <response code="401">You are not authorized</response>
+        /// <response code="500">Server error</response>
+        [HttpGet("my")]
+        [Authorize]
+        [ProducesResponseType(200)]
+        [ProducesResponseType(400)]
+        [ProducesResponseType(401)]
+        [ProducesResponseType(500)]
+        public async Task<IActionResult> GetEventsForCurrent([FromQuery] EventParameters eventParameters)
+        {
+            if (!eventParameters.ValidDateRange)
+                return BadRequest("Max date can't be less than min date.");
+
+            string userId = User.FindFirst(c => c.Type == "Id").Value;
+            var events = await _repository.Event.GetEventsForUserAsync(eventParameters, userId);
+
+            Response.Headers.Add("X-Pagination", JsonConvert.SerializeObject(events.MetaData));
+
+            var eventsDto = _mapper.Map<IEnumerable<EventDto>>(events);
+
+            return Ok(eventsDto);
+        }
+
+
         /// <summary>
         /// Create one new event
         /// </summary>
@@ -101,7 +131,7 @@ namespace Server.Controllers
         /// <response code="422">New event not valid</response>
         /// <response code="500">Server error</response>
         [HttpPost]
-        [Authorize(Roles = "Administrator")]
+        [Authorize]
         [ServiceFilter(typeof(ValidationFilterAttribute))]
         [ProducesResponseType(201)]
         [ProducesResponseType(400)]
@@ -112,6 +142,8 @@ namespace Server.Controllers
         public async Task<IActionResult> CreateEventAsync([FromBody] EventForCreationDto eventForCreationDto)
         {
             var eventEntity = _mapper.Map<Event>(eventForCreationDto);
+
+            eventEntity.CreaterId = User.FindFirst(c => c.Type == "Id").Value;
 
             _repository.Event.CreateEvent(eventEntity);
             await _repository.SaveAsync();
@@ -131,7 +163,7 @@ namespace Server.Controllers
         /// <response code="404">Event with this id not found</response>
         /// <response code="500">Server error</response>
         [HttpDelete("{id}")]
-        [Authorize(Roles = "Administrator")]
+        [Authorize]
         [ServiceFilter(typeof(ValidateEventExistsAttribute))]
         [ProducesResponseType(204)]
         [ProducesResponseType(401)]
@@ -141,6 +173,15 @@ namespace Server.Controllers
         public async Task<IActionResult> DeleteEventByIdAsync(Guid id)
         {
             var deletedEvent = HttpContext.Items["checkEvent"] as Event;
+
+            string userId = User.FindFirst(c => c.Type == "Id").Value;
+            bool isUserAdmin = User.IsInRole("Admin");
+
+            if (!deletedEvent.CreaterId.Equals(userId) && !isUserAdmin)
+            {
+                _logger.LogInfo($"User with this id {userId} has no rules to the event with id {deletedEvent.Id}");
+                return Forbid();
+            }
 
             _repository.Event.DeleteEvent(deletedEvent);
             await _repository.SaveAsync();
@@ -161,7 +202,7 @@ namespace Server.Controllers
         /// <response code="404">Event with this id not found</response>
         /// <response code="500">Server error</response>
         [HttpPut("{id}")]
-        [Authorize(Roles = "Administrator")]
+        [Authorize]
         [ServiceFilter(typeof(ValidationFilterAttribute))]
         [ServiceFilter(typeof(ValidateEventExistsAttribute))]
         [ProducesResponseType(204)]
@@ -174,6 +215,15 @@ namespace Server.Controllers
             [FromBody] EventForUpdateDto eventForUpdateDto)
         {
             var updateEvent = HttpContext.Items["checkEvent"] as Event;
+
+            string userId = User.FindFirst(c => c.Type == "Id").Value;
+            bool isUserAdmin = User.IsInRole("Admin");
+
+            if (!updateEvent.CreaterId.Equals(userId) && !isUserAdmin)
+            {
+                _logger.LogInfo($"User with this id {userId} has no rules to the event with id {updateEvent.Id}");
+                return Forbid();
+            }
 
             _mapper.Map(eventForUpdateDto, updateEvent);
             await _repository.SaveAsync();
